lib/workload/stateless/bs_runs_upload_manager/layers/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "bs_runs_upload_manager_tools"
version = "0.0.1"
description = "BSSH Manager Lambda Layers"
license = "GPL-3.0-or-later"
authors = [
    "Alexis Lucattini"
]
homepage = "https://github.com/umccr/orcabus"
repository = "https://github.com/umccr/orcabus"

[tool.poetry.dependencies]
python = "^3.11"
libica = "^2.3.0"
boto3 = "^1.28"
botocore = "^1.31"
verboselogs = "^1.7"
aws_requests_auth = "^0.4.3"
xmltodict = "^0.13.0"
pandas = "^2.2.0"
v2_samplesheet_maker = "^4.2.4"
wrapica = ">=1.0.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pyarrow = "^15.0.0"  # Pandas throws a warning if this is not installed
pytest = "^7.0.0"  # For testing only
# For typehinting only, not required at runtime
mypy-boto3-ssm = "^1.34"
mypy-boto3-secretsmanager = "^1.34"
mypy-boto3-stepfunctions = "^1.34"
mypy-boto3-s3 = "^1.34"